% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{bng_from_coords}
\alias{bng_from_coords}
\title{Convert eastings and northings to BNG grid reference}
\usage{
bng_from_coords(easting, northing, resolution)
}
\arguments{
\item{easting}{numeric vector of coordinates}

\item{northing}{numeric vector of coordinates}

\item{resolution}{numeric vector of resolutions in meters}
}
\value{
character vector of British National Grid references.
}
\description{
Internal helper function to create BNG reference strings.
}
\keyword{internal}
