% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traversal.R
\name{bng_kring}
\alias{bng_kring}
\alias{bng_kdisc}
\title{Spatial neighbourhoods}
\usage{
bng_kring(bng_ref, k, ...)

bng_kdisc(bng_ref, k, ...)
}
\arguments{
\item{bng_ref}{an object of type \code{BNGReference}}

\item{k}{numeric value measuring the number of grid squares traversed between
the ring and input BNG reference}

\item{...}{additional parameters. Not currently used}
}
\value{
an unordered collection of objects of type \code{BNGReference}.
}
\description{
Identify neighbours in a hollow ring or solid disc at grid distance 'k' from
a target BNG reference.
}
\details{
Additional details...
}
\examples{
# example code

}
