% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{bng_to_coords}
\alias{bng_to_coords}
\title{Convert BNG reference to eastings and northings}
\usage{
bng_to_coords(ref, position)
}
\arguments{
\item{ref}{BNG reference as a character vector}

\item{position}{character indicating which grid position to return}
}
\value{
matrix with two columns of easting and northings, respectively.
}
\description{
Internal helper function to create coordinate positions.
}
\details{
This internal function does not do extensive error checking/handling. The
calling functions need to handle inserting NAs for invalid references.
}
\keyword{internal}
