% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bng.R
\name{BNGReference}
\alias{BNGReference}
\alias{[.BNGReference}
\alias{[[.BNGReference}
\alias{[<-.BNGReference}
\alias{[[<-.BNGReference}
\alias{print.BNGReference}
\alias{c.BNGReference}
\alias{unique.BNGReference}
\title{BNG Reference objects}
\usage{
\method{[}{BNGReference}(x, i)

\method{[[}{BNGReference}(x, i)

\method{[}{BNGReference}(x, i) <- value

\method{[[}{BNGReference}(x, i) <- value

\method{print}{BNGReference}(x, ...)

\method{c}{BNGReference}(...)

\method{unique}{BNGReference}(x, incomparables = FALSE, ...)
}
\arguments{
\item{x}{object of class \code{BNGReference}.}

\item{i}{record selection.}

\item{value}{a suitable replacement value}
}
\description{
Functions to support working with objects of type
\code{BNGReference}.
}
\details{
The BNG is structured using a hierarchical system of grid squares at various
resolutions. At its highest level, the grid is divided into 100 km by 100 km
squares, each of which is identified by a two-letter code. Successive levels
of resolution further subdivide the grid squares into finer detail, down to
individual 1-meter squares.
}
\seealso{
\code{\link[=is_valid_bng]{is_valid_bng()}}, \code{\link[=as_bng_reference]{as_bng_reference()}}
}
