% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{geom_bng_intersects}
\alias{geom_bng_intersects}
\title{Generate a list of BNG indices for a geometry}
\usage{
geom_bng_intersects(geom, resolution)
}
\arguments{
\item{geom}{object of type \code{geos-geometry}.}

\item{resolution}{numeric value of the BNG grid resolution to use.}
}
\value{
list of where each element of the list corresponds to the input
geometry and contains a vector of BNG reference objects.
}
\description{
Using the bounding box, return the set of BNG tiles that intersect with the
geometry.
}
\keyword{internal}
