% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{geom_to_bng}
\alias{geom_to_bng}
\alias{geom_to_bng_intersection}
\alias{geom_to_bng.geos_geometry}
\alias{geom_to_bng.sf}
\alias{geom_to_bng_intersection.geos_geometry}
\alias{geom_to_bng_intersection.sf}
\title{Spatial index for geometries}
\usage{
geom_to_bng(geom, resolution, ...)

\method{geom_to_bng}{geos_geometry}(geom, resolution, ...)

\method{geom_to_bng}{sf}(geom, resolution, ...)

geom_to_bng_intersection(geom, resolution, ...)

\method{geom_to_bng_intersection}{geos_geometry}(geom, resolution, ...)

\method{geom_to_bng_intersection}{sf}(geom, resolution, ...)
}
\arguments{
\item{geom}{geometry object of type \code{geos-geometry} or \code{sf}}

\item{resolution}{spatial resolution of the BNG cell expressed in string or
integer values}
}
\description{
Generate the set of BNG cells that cover a geometry and optionally return an
intersection of the geometry with the grid.
}
\details{
Additional details...
}
