% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{is_valid_bng_resolution}
\alias{is_valid_bng_resolution}
\title{Validate BNG resolutions}
\usage{
is_valid_bng_resolution(resolution)
}
\arguments{
\item{resolution}{numeric or character vector of resolutions to test}
}
\value{
logical vector testing resolution
}
\description{
Helper function used to verify resolutions.
}
