% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osbng-package.R
\docType{package}
\name{osbng-package}
\alias{osbng}
\alias{osbng-package}
\title{\code{osbng}: Geospatial Grid Indexing and Interaction with the British National Grid}
\description{
Offers a streamlined programmatic interface to Ordnance Survey's British
National Grid (BNG) index system, enabling efficient spatial indexing and
analysis based on grid references. It supports a range of geospatial
applications, including statistical aggregation, data visualisation, and
interoperability across datasets. Designed for developers and analysts
working with geospatial data in Great Britain, \code{osbng} simplifies
integration with geospatial workflows and provides intuitive tools for
exploring the structure and logic of the BNG system.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ordnancesurvey.github.io/osbng-r/}
  \item \url{https://github.com/OrdnanceSurvey/osbng-r}
  \item Report bugs at \url{https://github.com/OrdnanceSurvey/osbng-r/issues}
}

}
\author{
\strong{Maintainer}: Chris Jochem \email{chris.jochem@os.uk}

Other contributors:
\itemize{
  \item Ordnance Survey Ltd [copyright holder, funder]
  \item Steve Kingston [contributor]
  \item Kate New [contributor]
  \item Tom Peterken [contributor]
}

}
\keyword{internal}
