% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bng.R
\name{print.BNGReference}
\alias{print.BNGReference}
\alias{pretty_format_bng}
\title{Printing BNG Reference Class}
\usage{
\method{print}{BNGReference}(x, spaces = TRUE, ...)

pretty_format_bng(x, spaces = TRUE)
}
\arguments{
\item{x}{an object of type \code{BNGReference}.}

\item{spaces}{boolean. Should standard spaces be added or removed for
"pretty" printing? Default is \code{TRUE} to add spaces.}

\item{...}{additional parameters}
}
\value{
\itemize{
\item \code{pretty_format_bng} inserts spaces and returns a vector of the references as strings.
\item \code{print} outputs the BNG references and invisibly returns the object.
}
}
\description{
Supporting formatting and printing of \code{BNGReference} objects.
}
\details{
Standard spaces are added: 1) after the two-letter prefix, 2) between
eastings and northings, and 3) before a quadrant suffix, when those
components exist in a grid reference.
}
\examples{
x <- as_bng_reference("SU1234")
pretty_format_bng(x)

pretty_format_bng(x, spaces = TRUE)

print(x)

print(x, spaces = TRUE)

}
