% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Regions}
\alias{Regions}
\title{Region Boundaries for England}
\format{
GeoPackage with one layer, 9 rows and 8 columns:
\describe{
\item{fid}{Row index.}
\item{RGN24CD}{Character. Region GSS code.}
\item{RGN24NM}{Character. Standardised Region name.}
\item{BNG_E}{Centroid eastings in British National Grid.}
\item{BNG_N}{Centroid northings in British National Grid.}
\item{LONG}{Centroid longitude coordinate.}
\item{LAT}{Centroid latitude coordinate.}
\item{GlobalID}{Database management.}
}
}
\source{
Office for National Statistics (ONS), released 16 February 2025,
Regions (December 2024) Boundaries EN BFC, \href{https://geoportal.statistics.gov.uk/}{ONS Open Geography Portal}.
}
\description{
This file (Regions_December_2024_Boundaries_EN_BFC.gpkg) contains the digital
vector boundaries for Regions in England as at December 2024. The boundaries
available are: (BFC) Full resolution - clipped to the coastline (Mean High
Water mark). The coordinate reference system is British National Grid
(OSGB36; EPSG:27700). Contains both Ordnance Survey and ONS Intellectual
Property Rights. Licenced under the Open Government Licence (OGLv3).
}
\keyword{data}
