% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{validate_bng_ref}
\alias{validate_bng_ref}
\title{Validate input}
\usage{
validate_bng_ref(x)
}
\arguments{
\item{x}{object to test}
}
\value{
\code{TRUE} when the input is a \code{BNGReference} object.
}
\description{
Internal helper function used to verify inputs.
}
\details{
Primarily called for the side-effect of stopping execution.
}
\keyword{internal}
