% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{validate_easting}
\alias{validate_easting}
\title{Validate positions}
\usage{
validate_easting(easting)
}
\arguments{
\item{easting}{object to test}
}
\value{
logical vector testing easting position
}
\description{
Internal helper function used to verify eastings.
}
\keyword{internal}
