% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{validate_northing}
\alias{validate_northing}
\title{Validate positions}
\usage{
validate_northing(northing)
}
\arguments{
\item{northing}{object to test}
}
\value{
logical vector testing northing position
}
\description{
Internal helper function used to verify northings.
}
\keyword{internal}
