% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traversal.R
\name{bng_distance}
\alias{bng_distance}
\alias{bng_dwithin}
\title{Distance calculations}
\usage{
bng_dwithin(bng_ref, d, ...)

bng_distance(bng_ref1, bng_ref2, by_element = FALSE)
}
\arguments{
\item{bng_ref}{object of class \code{BNGReference}.}

\item{d}{numeric. Distance expressed in metres.}

\item{...}{additional parameters. Not currently used.}

\item{bng_ref1, bng_ref2}{object of \code{BNGReference}}

\item{by_element}{logical. If \code{TRUE}, return a vector with distance
between each BNG reference. An error is raised if the \code{BNGReference}
objects are not the same length. If \code{FALSE}, return a dense matrix
with all pairwise distances.}
}
\value{
an unordered vector of \code{BNGReference} objects around a given
grid square within an absolute distance \code{d}.
}
\description{
Compute Euclidean distances between BNG references and distance-based
neighbours lists.
}
\examples{
# example code

# example code

}
