% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{bbox_to_bng}
\alias{bbox_to_bng}
\alias{bng_to_bbox}
\alias{bng_to_grid_geom}
\title{Convert bounding boxes}
\usage{
bbox_to_bng(...)

bng_to_bbox(bng_ref, ...)

bng_to_grid_geom(bng_ref, format = c("geos", "wkt", "sf"), ...)
}
\arguments{
\item{...}{additional parameters, not currently used}

\item{bng_ref}{vector of type \code{BNGReference} objects}

\item{format}{character indicating the type of geometry object to return}

\item{xmin, ymin, xmax, ymax}{numeric vector of bounding box coordinates}
}
\value{
\code{bng_to_bbox}: numeric vector of bounding coordinates. If
multiple references are supplied to \code{bng_ref} then a matrix of
coordinates is returned. \code{bbox_to_bng}: vector of type
\code{BNGReference} objects. \code{bng_to_grid_geom} converts the bounding
box coordinates into a Polygon using \code{geos}.
}
\description{
Create British National Grid reference from bounding boxes or convert grid
reference objects into bounding boxes.
}
\details{
Bounding boxes are expressed as four coordinates (min x, min y, max x, max
y). Coordinates must be in British National Grid projection (EPSG:27700).
These functions do not support coordinate transformations.
}
\examples{
# example code

# example code

# example code

}
