% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{bng_to_children}
\alias{bng_to_children}
\alias{bng_to_parent}
\title{Retrieve a hierarchy of BNG indices}
\usage{
bng_to_children(bng_ref, resolution, ...)

bng_to_parent(bng_ref, resolution, ...)
}
\arguments{
\item{bng_ref}{object of type \code{BNGReference}}

\item{resolution}{(optional) numeric value of the target resolution of
parent/child references. If omitted, the next "whole" resolution relative
to the input BNG reference is assumed.}

\item{...}{additional parameters. Not currently used}
}
\value{
child references will be a list of \code{BNGReference} objects with
each item in the list being the set of children for the input grid
reference. Parent references will be a vector of \code{BNGReference}
objects.
}
\description{
Identify the "parent" or "children" references which contain or are nested
within a given BNG reference.
}
\details{
More details
}
\examples{
# example code

}
