% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{geom_to_bng_intersection_explode}
\alias{geom_to_bng_intersection_explode}
\alias{geom_to_bng_intersection_explode.geos}
\alias{geom_to_bng_intersection_explode.sf}
\title{Spatial index data frame for geometries}
\usage{
geom_to_bng_intersection_explode(geom, resolution, add_index = FALSE, ...)

\method{geom_to_bng_intersection_explode}{geos}(geom, resolution, add_index = FALSE, ...)

\method{geom_to_bng_intersection_explode}{sf}(geom, resolution, add_index = FALSE, ...)
}
\arguments{
\item{geom}{geometry object of type \code{geos-geometry} or \code{sf}}

\item{resolution}{spatial resolution of the BNG cell expressed in string or
integer values}

\item{add_index}{logical. Should an index of the input feaure be added?}

\item{...}{additional parameters. Not currently used.}
}
\value{
a spatial data frame of type \code{sf} with the coordinate reference
system to British National Grid (EPSG:27700) and three columns for the
\code{BNGReference} object, the \code{is_core} property, and the
\code{geometry}. Optionally, a fourth column \code{index} will be added
when \code{add_index} is \code{TRUE}.
}
\description{
Generate a set of BNG Reference objects given a geometry and a specified
resolution. Specifically this function will provide results in a spatial data
frame format.
}
\details{
The BNG Reference objects returned represent the grid squares intersected by
the input geometry. This function followings the pattern of
\code{geom_to_bng_intersection()}, but flattens the list structure of results
into a spatial data frame. The \code{sf} package is required to use this
functionality.

When \code{add_index} is \code{TRUE}, the numeric index of the input geometry
will be added to the output data frame to indentify the source and enable
additional attributes to be rejoined to the results.
}
\seealso{
\code{\link[=geom_to_bng]{geom_to_bng()}}
}
