% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolution.R
\name{get_bng_resolution}
\alias{get_bng_resolution}
\alias{get_bng_resolution_string}
\title{BNG reference resolution}
\usage{
get_bng_resolution(bng_ref)

get_bng_resolution_string(bng_ref)
}
\arguments{
\item{bng_ref}{Vector of \code{BNGReference} objects to test.}
}
\value{
A vector of numeric values for \code{get_bng_resolution()} in metres
or character strings expressing the resolution of the grid references.
}
\description{
Find the spatial resolution (i.e. grid size) of a British National Grid
square.
}
\details{
How the resolution of grid reference is determined.
}
\examples{
get_bng_resolution(as_bng_reference("TQ1234"))

get_bng_resolution_string(as_bng_reference("TQ1234NE"))

}
