% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osbng-package.R
\docType{package}
\name{osbng-package}
\alias{osbng}
\alias{osbng-package}
\title{\code{osbng}: Access the British National Grid (BNG) index system}
\description{
\code{osbng} provides a programmatic interface to Ordnance Survey's British
National Grid (BNG) index system, supporting efficient grid-based spatial
indexing and analysis. This enables applications such as statistical
aggregation, data visualisation, and data interoperability using BNG grid
references. Designed for developers working with geospatial data in the
context of Great Britain, the library offers tools to simplify working with
the BNG, facilitating both technical integration into geospatial workflows
and practical exploration of the index system's structure.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ordnancesurvey.github.io/osbng-r/}
  \item \url{https://github.com/OrdnanceSurvey/osbng-r}
  \item Report bugs at \url{https://github.com/OrdnanceSurvey/osbng-r/issues}
}

}
\author{
\strong{Maintainer}: Chris Jochem \email{chris.jochem@os.uk}

Other contributors:
\itemize{
  \item Ordnance Survey Ltd [copyright holder, funder]
  \item Steve Kingston [contributor]
  \item Kate New [contributor]
  \item Tom Peterken [contributor]
}

}
\keyword{internal}
